+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Docker-compose : added option to disable automatic docker network attach to the carnotzet network when running inside a container.
 - DNS : added option to disable support for legacy DNS names
 - Runtime : added support for "exec"
+- Runtime : added support for specifying the shm_size of containers
 
 ## [1.7.7] - 2018-10-02
 ### Changed